Only save good images
This should fix #1, however, the code really needs to be cleaned up and
organized.

If we don't successfully download an image, we retry again two more
times. Hopefully this stops failed images from becoming successful. We
also now verify that the image is a valid JPEG image before moving on.